backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
paho-mqtt==1.6.1
redis==5.0.1
apscheduler==3.10.4
loguru==0.7.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
Pillow==10.1.0
matplotlib==3.8.2
pandas==2.1.3
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
